Minor renames variables group -> team, y extraigo dos funciones al utils
