+++ src/utils/getEntitiesUtils.js
@@ -7,3 +7,23 @@ export const getTeamById = (id, teams) => { // To-Do: usarla tmb en TopicTutor
     const group = teams?.find((g) => g.id === id);
     return group ? group : null;
 };
+
+ // Función para obtener el nombre del tutor por su id
+ export const getTutorNameById = (id, periodId, tutors) => {
+    const tutor = tutors.find(
+      (t) =>
+        t.tutor_periods &&
+        t.tutor_periods.some((tp) => tp.period_id === periodId && tp.id === id)
+    );
+
+    return tutor ? tutor.name + " " + tutor.last_name : "Sin asignar"; // Si no encuentra el tutor, mostrar 'Sin asignar'
+};
+
+export function formatDate(dateString) {
+    const date = new Date(dateString);
+    return date.toLocaleDateString("es-ES", {
+        day: "2-digit",
+        month: "2-digit",
+        year: "numeric",
+    });
+}
